backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
lightgbm==4.3.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
geojson==3.1.0
